MAJOR spring cleaning and reach functionality
