+++ Assets/LoganFolder/AIAttackState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d604f45fb9489504d8d410a7e682e7c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
