+++ Assets/LoganFolder/AIChaseState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0957d0e159c5e984b90dc8ab7e9e1611
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
