+++ Assets/LoganFolder/AIPatrolState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0f5699558d6be284a9e21ae7abd8d9f8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
