+++ Assets/LoganFolder/AITransitionState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b2a4f67865891e4a8dc536606ae1010
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
